dff is almost done (in trip at least) wip on eff
eff is a bit harde rand I need to be able to compute the p-th power of a
polynomial in Fp.

Then implem a lil mod p into sff in Fp or a sff in Z then mod p. have to
test/compute which one is faster.

Then hensel lift and we're good to implement for real in cpp